Move schemas to separate directory
